オーバーフローしていたので型を変更

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: tsito <tsito@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/30 22:12:10 by tsito             #+#    #+#             */
-/*   Updated: 2026/03/30 22:25:13 by tsito            ###   ########.fr       */
+/*   Updated: 2026/04/14 22:01:04 by tsito            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_atoi(const char *str)
 {
-	int	sign;
-	int	result;
+	int		sign;
+	long	result;
 
 	while (*str == ' ' || *str == '\t' || *str == '\n' || *str == '\v'
 		|| *str == '\f' || *str == '\r')
@@ -33,5 +33,5 @@ int	ft_atoi(const char *str)
 		result = result * 10 + *str - '0';
 		str++;
 	}
-	return (sign * result);
+	return ((int)(sign * result));
 }
